requirements.txt
langchain==0.0.184
openai==1.20.0
groq==0.10.0 
faiss-cpu==1.8.0
pandas==2.2.2
sentence-transformers==2.2.2
python-dotenv==0.21.0